feat: Add initial SearchHistory.module.css for styling
This commit introduces the 'SearchHistory.module.css' file,
which will house the styles for the SearchHistory component.
This file is currently empty but serves as the foundation for
defining the visual appearance and layout of the search history
feature.  It will allow us to encapsulate the component's styles
and maintain separation of concerns.